refactor: 抽离切换暗黑主题模式
