+++ src/App.vue
@@ -7,25 +7,19 @@
 </template>
 
 <script lang="ts" setup>
-  import { ConfigProvider, theme } from 'ant-design-vue';
+  import { ConfigProvider } from 'ant-design-vue';
   import { AppProvider } from '@/components/Application';
   import { useTitle } from '@/hooks/web/useTitle';
   import { useLocale } from '@/locales/useLocale';
-  import { useRootSetting } from '/@/hooks/setting/useRootSetting';
-  import 'dayjs/locale/zh-cn';
-  import { computed } from 'vue';
-  import { ThemeEnum } from '/@/enums/appEnum';
 
-  const { darkAlgorithm } = theme;
+  import 'dayjs/locale/zh-cn';
+  import { useDarkModeTheme } from '@/hooks/setting/useDarkModeTheme';
+
   // support Multi-language
   const { getAntdLocale } = useLocale();
-  const { getDarkMode } = useRootSetting();
 
-  const isDark = computed(() => getDarkMode.value === ThemeEnum.DARK);
+  const { isDark, darkTheme } = useDarkModeTheme();
 
-  const darkTheme = {
-    algorithm: [darkAlgorithm],
-  };
   // Listening to page changes and dynamically changing site titles
   useTitle();
 </script>
